+++ code_pylint.py
@@ -29,16 +29,16 @@ MAX_ERROR_BY_TYPE = {
     "protected-access": 48,  # Highly dependant on our "private" conventions. Keeps getting raised
     "arguments-differ": 1,
     "too-many-locals": 5,  # Reduce by dropping vectored objects
-    "too-many-branches": 9,  # Huge refactor needed. Will be reduced by schema refactor
+    "too-many-branches": 8,  # Huge refactor needed. Will be reduced by schema refactor
     "unused-argument": 3,  # Some abstract functions have unused arguments (plot_data). Hence cannot decrease
     "cyclic-import": 2,  # Still work to do on Specific based DessiaObject
     "too-many-arguments": 18,  # Huge refactor needed
-    "too-few-public-methods": 3,  # Abstract classes (Errors, Checks,...)
-    "too-many-return-statements": 8,  # Huge refactor needed. Will be reduced by schema refactor
+    "too-few-public-methods": 4,  # Abstract classes (Errors, Checks,...)
+    "too-many-return-statements": 7,  # Huge refactor needed. Will be reduced by schema refactor
     "import-outside-toplevel": 5,  # TODO : will reduced in a future work (when tests are ready)
     "too-many-instance-attributes": 6,  # Huge refactor needed (workflow, etc...)
     "broad-exception-caught": 9,  # Necessary in order not to raise non critical errors.
-    "too-many-public-methods": 2,  # Try to lower by splitting DessiaObject and Workflow
+    "too-many-public-methods": 3,  # Try to lower by splitting DessiaObject and Workflow
 }
 
 ERRORS_WITHOUT_TIME_DECREASE = ['protected-access', 'arguments-differ', 'too-many-locals', 'too-many-branches',