+++ CHANGELOG.md
@@ -5,14 +5,75 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.17.0
+
+### Added
+
+- Schemas : Add a new type to compute string literals
+- Workflow : Display by default is now tasks only
+
+
+## 0.16.5
+
+### HOTFIX
+
+- Workflow : Reset NBVs type to last downstream one found while regenerating workflow pipes from deserialization 
+
+
+## 0.16.4
+
+### Fix
+
+- Workflow : Stop deserializing and setting block Variable. Instead, reconfigure them based on incoming dict
+- Workflow : Blocks are not generating their to_dict with pointers anymore
+
+
+### Feature
+
+- Workflow : Display by default is now tasks only
+
+
+## 0.16.3
+
+### Fix
+
+- Dataset : Update plot to new decorator paradigm
+- Workflow : Reset variables types for block defined from user code (should not trust frontend json on variable types)
+- Workflow : If type is provided, Variables do not reset their type based on their default value anymore
+
+### Refactor
+
+- Workflow : Use AttributeType to compute variable type for ModelAttribute Blocks
+
+
+## 0.16.2
+
+### Fix
+
+- Export: Add block_index if export blocks
+- PhysicalObject : Remove to_stl encoding element
+- Schema : Take `None` default value into account
+- WorkflowRun : Compute memory usage, subtract before_memory
+
+
+### Build
+
+- Numpy : Restrict version to lower 2.0.0
+
 
 ## 0.16.1
 
 ### Added
 
-- Export : Define export method with decorator 
+- Export : Define export method with decorator
 
-### Performance
+### Fix
+
+- ModelMethod : Call the right dict_to_object of MethodType
+- MultiPlot : Add selector_name MultiPlot script
+- WorkflowRun : Add markdown to display_settings
+
+### Tests
 
 - Schema : Add new unittest to test get_import_names methods